news/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
httpx==0.27.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary
python-dotenv==1.0.1
pydantic==2.7.3
redis==5.0.4